+++ src/frontend/cards/insertable-card.tsx
@@ -22,10 +22,7 @@ import {
     FavoriteInsertableItem
 } from "../favorites/favorite-button";
 import { useIsInsertableHidden, useSetVisibilityMutation } from "./card-hooks";
-import {
-    getInsertableBuildIssues,
-    getInsertableStateRows
-} from "./build-status";
+import { InsertableStatusBadge } from "./build-status";
 import {
     AdminOptionsSubmenu,
     CardTitle,
@@ -66,7 +63,6 @@ export function InsertableCard(props: InsertableCardProps): ReactNode {
     const favorites = useFavoritesQuery().data?.favorites;
 
     const isHidden = useIsInsertableHidden(insertable);
-    const buildIssues = getInsertableBuildIssues(insertable);
 
     const isAssemblyInPartStudio = useIsAssemblyInPartStudio(
         insertable.elementType
@@ -99,8 +95,9 @@ export function InsertableCard(props: InsertableCardProps): ReactNode {
                     title={insertable.name}
                     thumbnailUrls={insertable.thumbnailUrls}
                     showHiddenTag={!insertable.isVisible}
-                    buildIssues={buildIssues}
-                    buildStateRows={getInsertableStateRows(insertable)}
+                    buildStatusBadge={
+                        <InsertableStatusBadge insertable={insertable} />
+                    }
                 />
             }
             rightSection={
