+++ src/frontend/groups/group-card.tsx
@@ -25,7 +25,7 @@ import {
     ReloadThumbnailMenuItem
 } from "../cards/card-components";
 import { AddGroupItem } from "./add-group-menu";
-import { getGroupStateRows, useGroupBuildIssues } from "../cards/build-status";
+import { GroupStatusBadge } from "../cards/build-status";
 import {
     libraryQueryKey,
     libraryQueryMatchKey,
@@ -45,8 +45,6 @@ interface GroupCardProps extends PropsWithChildren {
 export function GroupCard(props: GroupCardProps): ReactNode {
     const { group } = props;
     const navigate = useNavigate();
-    const buildIssues = useGroupBuildIssues(group);
-
     return (
         <ItemRow
             onClick={() => {
@@ -59,8 +57,7 @@ export function GroupCard(props: GroupCardProps): ReactNode {
                 <CardTitle
                     title={group.name}
                     thumbnailUrls={group.thumbnailUrls}
-                    buildIssues={buildIssues}
-                    buildStateRows={getGroupStateRows(group)}
+                    buildStatusBadge={<GroupStatusBadge group={group} />}
                 />
             }
             rightSection={<IconArrowRight size={IconSize.SMALL} />}
